requirements.txt
fastapi
uvicorn
pyannote.audio
whisper
ffmpeg-python
demucs
torch
